fi-nexus-backend-service/requirements.txt
google-adk==1.7.0
